feat: :sparkles: Input box backend for searching
